--- DrawableConsole/BoardPiece.cpp
@@ -1,2 +1,0 @@
-#include "pch.h"
-#include "BoardPiece.h"
