+++ DrawableConsole/Concerto.h
@@ -5,20 +5,20 @@
 #include <math.h>
 #include <time.h>
 #include "Vector2.h"
-#include "Board.h"
+#include "SYDEOldCodeBoard.h"
 #include <stdlib.h> 
 #include <string>
 #include <windows.h>
 #include <mmsystem.h>
 #include <conio.h>
 #include <fstream>
-#include "AI.h"
+#include "SYDEOldCodeAI.h"
 #include "Settings.h"
 #include "PhysicsObject.h"
 #include "RigidBody.h"
 #include "BackgroundObj.h"
 #include "Background.h"
-#include "Characters.h"
+#include "SYDEOldCodeCharacters.h"
 #include "Artwork.h"
 #include "AssetsClass.h"
 #include "CustomAsset.h"
