+++ DrawableConsole/DrawableConsole.rc
@@ -61,8 +61,8 @@ LANGUAGE LANG_ENGLISH, SUBLANG_ENGLISH_AUS
 //
 
 VS_VERSION_INFO VERSIONINFO
- FILEVERSION 1,2,7,0
- PRODUCTVERSION 1,2,7,0
+ FILEVERSION 1,2,8,0
+ PRODUCTVERSION 1,2,8,0
  FILEFLAGSMASK 0x3fL
 #ifdef _DEBUG
  FILEFLAGS 0x1L
@@ -79,12 +79,12 @@ BEGIN
         BEGIN
             VALUE "CompanyName", "TODO: <Company name>"
             VALUE "FileDescription", "TODO: <File description>"
-            VALUE "FileVersion", "1.2.7.0"
+            VALUE "FileVersion", "1.2.8.0"
             VALUE "InternalName", "Drawable.exe"
             VALUE "LegalCopyright", "Copyright (C) 2020"
             VALUE "OriginalFilename", "Drawable.exe"
             VALUE "ProductName", "SYDE Engine"
-            VALUE "ProductVersion", "1.2.7.0"
+            VALUE "ProductVersion", "1.2.8.0"
         END
     END
     BLOCK "VarFileInfo"
