+++ DrawableConsole/SYDEEngineOldCode.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "SYDEOldCodeAI.h"
+#include "SYDEOldCodeBoard.h"
+#include "SYDEOldCodeBoardPiece.h"
+#include "SYDEOldCodeCharacters.h"
