+++ DrawableConsole/SYDEOldCodeBoardPiece.cpp
@@ -1,0 +1,2 @@
+#include "pch.h"
+#include "SYDEOldCodeBoardPiece.h"
