+++ Readme.md
@@ -125,6 +125,10 @@ Inside our main function, we can initialize all of our variables and call up SYD
 
 ## Release Notes
 
+### 1.2.8.0 18/05/2021
+- Added Comments
+- Separated outdated headers into SYDEOldCode.h
+
 ### 1.2.7.0 18/05/2021
 - Added SYDESound.h
 - get_bmp_as_array fixes issue of bmp loading
