+++ SYDESoundDemo/SYDESoundDemoMain.cpp
@@ -17,7 +17,6 @@ Settings config("EngineFiles\\Settings\\configSettings.sc");
 ConsoleWindow window(windowHeight);
 //DRAWING SETTINGS
 BackgroundClass bgVars;
-Characters charVars;
 Artwork artVars;
 //OTHER ASSETS
 AssetsClass astVars;
